requirements.txt
anthropic>=0.40.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
